add xdebug to allow code-coverage
Installed and enabled xdebug so that we can have code-coverage in ci. This required us to install some extras like autoconf, g++, and make, but we then realized that we could purge a few extra packages after the container build was in a usable state.

Size increased from ~52MB to ~117MB, I believe; must be done though.